Added ft_fdnstr.c, ft_fdstr.c, ft_printbits.c, ft_strncpy_u.c, ft_strnjoin.c, ft_strupper.c, ft_swap_endian32.c, ft_swap_endian64.c && added 'S' flag to ft_printf

+++ ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hbhuiyan <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hbhuiyan <hbhuiyan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/22 14:51:32 by hbhuiyan          #+#    #+#             */
-/*   Updated: 2018/10/22 14:52:53 by hbhuiyan         ###   ########.fr       */
+/*   Updated: 2019/08/12 13:44:32 by hbhuiyan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+uintmax_t	ft_strlen(const char *str)
 {
-	size_t i;
+	uintmax_t i;
 
 	i = 0;
 	while (str[i])
